requirements.txt
dragonfly2>=0.34.0
wxpython
pillow>=6.2.2
tomlkit>=0.11.8
future>=0.18.2
mock>=3.0.5
appdirs>=1.4.3
scandir>=1.10.0
pyvda==0.0.8
PySide6
six
openai
sounddevice
numpy
soundfile
pyperclip
webrtcvad-wheels # Optional, but recommended for VAD implementation
